# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Application Management API
  version: 1.0.0
  description: API for application submission and review workflows.
paths:
...
  /api/v1/auth/password/reset/:
    post:
      operationId: auth_password_reset_request
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestRequest'
        required: true
      security:
      - {}
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
  /api/v1/auth/password/reset/{uid}/{token}/:
    post:
      operationId: auth_password_reset_confirm
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      - in: path
        name: uid
        schema:
          type: string
        required: true
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequestRequest'
        required: true
      security:
      - {}
      responses:
        '204':
          description: No response body
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: ''
...
components:
  schemas:
...
    Detail:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
    Error:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        errors:
          type: object
          additionalProperties: {}
          nullable: true
      required:
      - code
      - detail
      - errors
...
    PasswordResetConfirmRequestRequest:
      type: object
      description: Public request body; reset credentials are supplied by the URL.
      properties:
        new_password:
          type: string
          writeOnly: true
          minLength: 1
        new_password_confirm:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - new_password
      - new_password_confirm
    PasswordResetRequestRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
      required:
      - email
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid